fix : 코드 재수정
