+++ src/App.js
@@ -1,9 +1,9 @@
-
 import "./App.css";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import Home from "./pages/home/Home";
 import GroupGenerator from "./pages/groupGenrator/GroupGenerator";
 import MissionImagesUpload from "./pages/MissionImagesUpload/MissionImagesUpload";
+import Mission from "./pages/Missions/Mission";
 
 function App() {
   return (
@@ -12,10 +12,8 @@ function App() {
         <Routes>
           <Route path="/" Component={Home}></Route>
           <Route path="/GroupGenerator" Component={GroupGenerator}></Route>
-          <Route
-            path="/MissionImagesUpload"
-            Component={MissionImagesUpload}
-          ></Route>
+          <Route path="/MissionImagesUpload" Component={MissionImagesUpload}></Route>
+          <Route path="Mission" Component={Mission}></Route>
         </Routes>
       </BrowserRouter>
     </div>
