+++ src/pages/Missions/Mission.css
@@ -26,11 +26,11 @@
 
 .change-container {
   height: 20px;
-  width: 130px;
+  width: 200px;
   border: none;
   display: flex;
   align-items: center;
-  margin-left: 66%;
+  margin-left: 70%;
   float: right;
 }
 
@@ -58,6 +58,8 @@
 .changeButton {
   border: none;
   background-color: white;
+  width: 20px;
+  height: 20px;
 }
 
 .question {
